smt-runner/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
docker==2.4.0
docker-pycreds==0.2.1
filemagic==1.6
idna==2.5
jsonschema==2.5.1
matplotlib==2.0.2
numa==1.4.4
numpy==1.15.4
psutil==4.3.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2017.2
PyYAML==3.12
requests==2.18.3
six==1.10.0
urllib3==1.22
websocket-client==0.44.0
